src/main/java/ru/netology/smartHome/Radio.java
package ru.netology.smartHome;

public class Radio {
    private int minStation = 0;
    private int maxStation = 9;

    private int currentStation;

    private int minVolume = 0;
    private int maxVolume = 100;

    private int currentVolume;

    public Radio () {
        this.maxStation = maxStation;
        this.currentStation = currentStation;
    }

    public Radio (int size) {
        if (size > 10) {
            this.maxStation = maxStation;
        } else {
            this.maxStation = size - 1;
        }
        this.currentStation = currentStation;
    }

    public int getCurrentStation() {
        return currentStation;
    }

    public int getMaxStation() {
        return maxStation;
    }

    public int getMinStation() {
        return minStation;
    }

    public int getCurrentVolume() {
        return currentVolume;
    }

    public int getMaxVolume() {
        return maxVolume;
    }

    public int getMinVolume() {
        return minVolume;
    }

    public void setCurrentStation(int newCurrentStation) {
        if (newCurrentStation < minStation) {
            return;
        }
        if (newCurrentStation > maxStation) {
            return;
        }
        currentStation = newCurrentStation;
    }

    public void next() {
        if (currentStation + 1 > maxStation) {
            currentStation = 0;
        } else {
            currentStation++;
        }
    }

    public void prev() {
        if (currentStation - 1 < minStation) {
            currentStation = 9;
        } else {
            currentStation--;
        }
    }


    public void setCurrentVolume(int newCurrentVolume) {
        if (newCurrentVolume < minVolume) {
            return;
        }
        if (newCurrentVolume > maxVolume) {
            return;
        }
        currentVolume = newCurrentVolume;
    }


    public void decreaseVolume() {
        if (currentVolume > minVolume) {
            currentVolume = currentVolume - 1;
        }
    }

    public void increaseVolume() {
        if (currentVolume < maxVolume) {
            currentVolume = currentVolume + 1;
        }
    }
}
